// src/employee-module/employee/dtos/CreateEmpAddress.dto.ts
export class CreateEmpAddress {
  empId: number;
  addressType: string;
  addressDetail: string;
  subDistrict: string;
  district: string;
  province: string;
  postalCode: string;
  country: string;
  phone: string;
  email: string;
  latitude: number;
  longitude: number;
  createdBy: number;
  createdDate: Date;
  modifiedBy: number;
  modifiedDate: Date;
  isDeleted: boolean;
  userId: number;
  companyId: number;
}